backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
django-cors-headers>=4.3
django-filter>=24.0
python-decouple>=3.8
reportlab>=4.1
openpyxl>=3.1

# Testing
pytest>=7.4
pytest-django>=4.5
pytest-cov>=4.1
pytest-mock>=3.11
factory-boy>=3.3
